# api/docs/swagger.yaml
definitions:
  models.Basket:
    properties:
      created_at:
        type: string
      customer_id:
        type: string
      id:
        type: string
      total_sum:
        type: integer
      updated_at:
        type: string
    type: object
  models.BasketProduct:
    properties:
      basket_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      product_id:
        type: string
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
  models.BasketProductResponse:
    properties:
      basketProducts:
        items:
          $ref: '#/definitions/models.BasketProduct'
        type: array
      count:
        type: integer
    type: object
  models.BasketResponse:
    properties:
      baskets:
        items:
          $ref: '#/definitions/models.Basket'
        type: array
      count:
        type: integer
    type: object
...
  models.CreateBasket:
    properties:
      customer_id:
        type: string
      total_sum:
        type: integer
    type: object
  models.CreateBasketProduct:
    properties:
      basket_id:
        type: string
      product_id:
        type: string
      quantity:
        type: integer
    type: object
...
  models.CustomerLoginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
...
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
...
  models.UpdateBasket:
    properties:
      customer_id:
        type: string
      total_sum:
        type: integer
    type: object
  models.UpdateBasketProduct:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
  /auth/customer/login:
    post:
      consumes:
      - application/json
      description: Customer login
      parameters:
      - description: login
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.CustomerLoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Basket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Customer login
      tags:
      - auth
  /basket:
    post:
      consumes:
      - application/json
      description: create a new basket
      parameters:
      - description: basket
        in: body
        name: basket
        schema:
          $ref: '#/definitions/models.CreateBasket'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Basket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Creates a new basket
      tags:
      - basket
  /basket/{id}:
    delete:
      consumes:
      - application/json
      description: delete basket
      parameters:
      - description: basket_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete basket
      tags:
      - basket
    get:
      consumes:
      - application/json
      description: get basket by id
      parameters:
      - description: basket_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Basket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get basket by id
      tags:
      - basket
    put:
      consumes:
      - application/json
      description: update basket
      parameters:
      - description: basket_id
        in: path
        name: id
        required: true
        type: string
      - description: basket
        in: body
        name: basket
        schema:
          $ref: '#/definitions/models.UpdateBasket'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Basket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update basket
      tags:
      - basket
  /basketProduct:
    post:
      consumes:
      - application/json
      description: create a new basketProduct
      parameters:
      - description: basket
        in: body
        name: basket
        schema:
          $ref: '#/definitions/models.CreateBasketProduct'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BasketProduct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Creates a new basketProduct
      tags:
      - basketProduct
  /basketProduct/{id}:
    delete:
      consumes:
      - application/json
      description: delete basketProduct
      parameters:
      - description: basketProduct_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete basketProduct
      tags:
      - basketProduct
    get:
      consumes:
      - application/json
      description: get basketProduct by id
      parameters:
      - description: basketProduct_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BasketProduct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get basketProduct by id
      tags:
      - basketProduct
    put:
      consumes:
      - application/json
      description: update basketProduct
      parameters:
      - description: basketProduct_id
        in: path
        name: id
        required: true
        type: string
      - description: basketProduct
        in: body
        name: basketProduct
        schema:
          $ref: '#/definitions/models.UpdateBasketProduct'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BasketProduct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update basketProduct
      tags:
      - basketProduct
  /basketProducts:
    get:
      consumes:
      - application/json
      description: get basket list
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      - description: basket_id
        in: query
        name: basket_id
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BasketProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get basket list
      tags:
      - basketProduct
  /baskets:
    get:
      consumes:
      - application/json
      description: get basket list
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BasketResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get basket list
      tags:
      - basket
...
swagger: "2.0"